CGTrader_SUBMIT_SERVER/requirements.txt
# CGTrader Submit Server Dependencies
# Optimized for low memory usage

# Web framework
flask>=3.0.0
gunicorn>=21.0.0

# Browser automation
selenium>=4.15.0
webdriver-manager>=4.0.0

# Telegram bot
python-telegram-bot>=20.7

# OpenAI Vision API
openai>=1.6.0

# HTTP client
httpx>=0.25.0
requests>=2.31.0

# Database
aiosqlite>=0.19.0

# Utilities
python-dotenv>=1.0.0

# HTML parsing
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Cloudflare bypass
cloudscraper>=1.2.0
